Add Encoding/Decoded
These helpers make it convenient to convert a plain AsyncRead/AsyncWrite
to a Stream/Sink, rather than forcing you to ues a combined object as
with Codec. This improves usability of unidirectional byte streams such
as Unix stdio and pipes, and enables two-part streams like TCP and UDP
sockets to be more conveniently handled as two separate parts.