Add a required AsyncWrite::shutdown method
This commit adds a method for `AsyncWrite` streams to hook into shutdown
semantics and otherwise perform clean shutdowns and such. A method
`AsyncWrite::shutdown` is added to embody these shutdown semantics,
returning `Ready` when it's done and `NotReady` if it's scheduled to do more
work. Currently it's also documented as basically invalidating the rest of the
I/O object, only allowing future `shutdown` method calls.

Additionally a `io::shutdown` future was added to shut down in a future-style
fashion.

Closes #11